Dietician

+++ database/migrations/2024_04_28_045849_create_dieticians_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('dieticians', function (Blueprint $table) {
             $table->id();
-            $table->integer('login_id');
+            $table->unsignedBigInteger('login_id');
             $table->string('name');
             $table->string('email');
             $table->string('address');
